requirements-gcp.txt
# Core Framework
langchain==0.1.0
langchain-google-genai==0.0.6

# GCP Services
google-cloud-aiplatform==1.38.1
google-cloud-firestore==2.13.1
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.16.4
vertexai==1.38.1

# Vector & Embeddings
sentence-transformers==2.2.2

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.16.0
requests==2.31.0
lxml==5.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
PyPDF2==3.0.1
python-docx==1.1.0

# UI
streamlit==1.29.0
plotly==5.18.0

# PDF Generation
fpdf==1.7.2
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1